# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/JoinGameSelectCamp": {
            "post": {
                "description": "选择阵营-新进入用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "魔塔"
                ],
                "summary": "选择阵营-新进入用户",
                "parameters": [
                    {
                        "description": "选择阵营-新进入用户",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MTJoinGameSelectCampReq"
                        }
                    }
                ],
                "responses": {
                    "1": {
                        "description": "选择阵营-新进入用户",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.MTJoinGameSelectCampResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
        "/betList": {
            "post": {
                "description": "魔塔手动押注列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "魔塔"
                ],
                "summary": "魔塔手动押注列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户姓名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "魔塔手动押注列表",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MtBetReq"
                        }
                    }
                ],
                "responses": {
                    "1": {
                        "description": "魔塔手动押注列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.MtBetListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
        "/betOnAutoNum": {
            "post": {
                "description": "自动用户:充值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "魔塔"
                ],
                "summary": "自动用户:充值",
                "parameters": [
                    {
                        "description": "自动用户:充值",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MTBetOnAutoNumReq"
                        }
                    }
                ],
                "responses": {
                    "1": {
                        "description": "自动用户:充值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.MTBetOnAutoNumResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
        "/betOnNum": {
            "post": {
                "description": "手动押注：充值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "魔塔"
                ],
                "summary": "手动押注：充值",
                "parameters": [
                    {
                        "description": "手动押注：充值",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MTBetOnNumReq"
                        }
                    }
                ],
                "responses": {
                    "1": {
                        "description": "手动押注：充值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.MTBetOnNumResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
...
        "/isAutoUser": {
            "post": {
                "description": "是否成为自动用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "魔塔"
                ],
                "summary": "是否成为自动用户",
                "parameters": [
                    {
                        "description": "是否成为自动用户",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MTIsAutoReq"
                        }
                    }
                ],
                "responses": {
                    "1": {
                        "description": "是否成为自动用户",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.MTIsAutoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
...
        "/mtCurrStatus": {
            "post": {
                "description": "魔塔当前状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "魔塔"
                ],
                "summary": "魔塔当前状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户姓名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "魔塔当前状态",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MTStatusReq"
                        }
                    }
                ],
                "responses": {
                    "1": {
                        "description": "魔塔当前状态",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.MTStatusResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/common.JSONResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "models.BetList": {
            "type": "object",
            "properties": {
                "bet": {
                    "description": "押注",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "models.MTBetOnAutoNumReq": {
            "type": "object",
            "properties": {
                "bet": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "models.MTBetOnAutoNumResp": {
            "type": "object"
        },
        "models.MTBetOnNumReq": {
            "type": "object",
            "properties": {
                "bet": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "models.MTBetOnNumResp": {
            "type": "object"
        },
        "models.MTIsAutoReq": {
            "type": "object",
            "properties": {
                "is_auto": {
                    "description": "0:取消 1:成为自动玩家",
                    "type": "integer"
                }
            }
        },
        "models.MTIsAutoResp": {
            "type": "object"
        },
        "models.MTJoinGameSelectCampReq": {
            "type": "object",
            "properties": {
                "camp": {
                    "description": "1=魔王军 2=圣盟军",
                    "type": "integer"
                }
            }
        },
        "models.MTJoinGameSelectCampResp": {
            "type": "object"
        },
...
        "models.MTStatusReq": {
            "type": "object"
        },
...
        "models.MtBetListResp": {
            "type": "object",
            "properties": {
                "bet_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BetList"
                    }
                }
            }
        },
        "models.MtBetReq": {
            "type": "object"
        },
...
    }
}